LF-4163 Integrate container into form framework
Replaces dummy container, removes progress bar, and fixes form field array value. Because of tsc complaints, I am committing in one go.
